requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
joblib
sqlalchemy
geoalchemy2
psycopg2-binary
python-dotenv
pydantic